Removed (by explicitly not using variables and so on) all warnings in th qcodeedit-module. Only a moc warning is still showing up.
git-svn-id: svn://svn.code.sf.net/p/texstudio/code/trunk@159 92cce386-e96d-4106-809f-596c673a3fd8

+++ qcodeedit/lib/qreliablefilewatch.cpp
@@ -3,7 +3,7 @@
 ** Copyright (C) 2006-2009 fullmetalcoder <fullmetalcoder@hotmail.fr>
 **
 ** This file is part of the Edyuk project <http://edyuk.org>
-** 
+**
 ** This file may be used under the terms of the GNU General Public License
 ** version 3 as published by the Free Software Foundation and appearing in the
 ** file GPL.txt included in the packaging of this file.
@@ -33,31 +33,31 @@ QReliableFileWatch::QReliableFileWatch(QObject *p)
 {
 	connect(this, SIGNAL( fileChanged(QString) ),
 			this, SLOT  ( sourceChanged(QString) ) );
-	
+
 }
 
 QReliableFileWatch::~QReliableFileWatch()
 {
-	
+
 }
 
 void QReliableFileWatch::addWatch(const QString& file, QObject *recipient)
 {
 	QHash<QString, Watch>::iterator it = m_targets.find(file);
-	
+
 	if ( it != m_targets.end() )
 	{
 		it->recipients << recipient;
 	} else {
 		QFile f(file);
-		
+
 		Watch w;
 		w.state = Clean;
 		w.checksum = -1;
 		w.size = f.size();
 		w.recipients << recipient;
 		m_targets[file] = w;
-		
+
 		addPath(file);
 	}
 }
@@ -70,20 +70,20 @@ void QReliableFileWatch::removeWatch(QObject *recipient)
 void QReliableFileWatch::removeWatch(const QString& file, QObject *recipient)
 {
 	QHash<QString, Watch>::iterator it = m_targets.find(file);
-	
+
 	if ( it == m_targets.end() )
 	{
 		if ( !recipient )
 			return;
-		
+
 		// given recipient stop watching any file
-		
+
 		it = m_targets.begin();
-		
+
 		while ( it != m_targets.end() )
 		{
 			int n = it->recipients.removeAll(recipient);
-			
+
 			if ( n && it->recipients.isEmpty() )
 			{
 				// no more recipients watching this file
@@ -98,7 +98,7 @@ void QReliableFileWatch::removeWatch(const QString& file, QObject *recipient)
 		{
 			// given recipient stops watching given file
 			it->recipients.removeAll(recipient);
-			
+
 			if ( it->recipients.isEmpty() )
 			{
 				// no more recipients watching this file
@@ -109,17 +109,17 @@ void QReliableFileWatch::removeWatch(const QString& file, QObject *recipient)
 			// stop watching given file at all
 			m_targets.erase(it);
 		}
-	} 
+	}
 }
 
 void QReliableFileWatch::timerEvent(QTimerEvent *e)
 {
 	if ( e->timerId() != m_timer.timerId() )
 		return QFileSystemWatcher::timerEvent(e);
-	
+
 	int postponedEmissions = 0;
 	QHash<QString, Watch>::iterator it = m_targets.begin();
-	
+
 	while ( it != m_targets.end() )
 	{
 		if ( it->state & Duplicate )
@@ -130,28 +130,28 @@ void QReliableFileWatch::timerEvent(QTimerEvent *e)
 		} else if ( it->state & Recent ) {
 			// send signal
 			it->state = Clean;
-			
+
 			QFile f(it.key());
-			
-			if ( f.size() == it->size )
+
+			if ( f.size() == (unsigned int)(it->size) )
 			{
 				// TODO : avoid signal emission if checksum match
 				// DO this in editor or here?
 			}
-			
+
 			//qDebug("%s emission.", qPrintable(it.key()));
-			
+
 			it->recipients.removeAll(0);
-			
+
 			foreach ( QObject *r, it->recipients )
 				QMetaObject::invokeMethod(r, "fileChanged", Q_ARG(QString, it.key()));
-			
+
 			//it = m_state.erase(it);
 		}
-		
+
 		++it;
 	}
-	
+
 	if ( postponedEmissions )
 	{
 		//qDebug("%i postponed emissions", postponedEmissions);
@@ -162,20 +162,20 @@ void QReliableFileWatch::timerEvent(QTimerEvent *e)
 void QReliableFileWatch::sourceChanged(const QString& filepath)
 {
 	m_timer.stop();
-	
+
 	QHash<QString, Watch>::iterator it = m_targets.find(filepath);
-	
+
 	if ( it == m_targets.end() )
 		return;
-	
+
 	//qDebug("%s modified.", qPrintable(filepath));
-	
+
 	if ( it->state )
 	{
 		it->state = Recent | Duplicate;
 	} else {
 		it->state = Recent;
 	}
-	
+
 	m_timer.start(20, this);
 }
